refactor: HomeContentAdapter의 viewType 지정방식 수정
