+++ app/src/main/java/woowacourse/movie/home/presenter/MovieHomePresenter.kt
@@ -1,6 +1,5 @@
 package woowacourse.movie.home.presenter
 
-import android.util.Log
 import woowacourse.movie.data.MovieRepository.getAllAdvertisements
 import woowacourse.movie.data.MovieRepository.getAllMovies
 import woowacourse.movie.home.presenter.contract.MovieHomeContract
@@ -22,8 +21,6 @@ class MovieHomePresenter(private val movieHomeContractView: MovieHomeContract.Vi
             }
             homeContents.add(movie)
         }
-
-        Log.e("a", homeContents.toString())
 
         movieHomeContractView.displayHomeContents(homeContents)
     }
