+++ app/src/test/java/woowacourse/movie/home/presenter/TheaterSelectionPresenterTest.kt
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test
 import woowacourse.movie.R
 import woowacourse.movie.data.MovieRepository
 import woowacourse.movie.home.presenter.contract.TheaterSelectionContract
-import woowacourse.movie.model.Movie
+import woowacourse.movie.home.view.adapter.movie.HomeContent.*
 import woowacourse.movie.model.MovieDate
 import woowacourse.movie.model.Theater
 import java.time.LocalDate
